Extract manager order projections
