+++ docs/white-label-platform-plan.md
@@ -19,6 +19,8 @@ foundation-слоя; действующий MVN остаётся первым и
   межтабличная согласованность закреплены contract-миграцией.
 - До второго внутреннего storefront остаются доверенное разрешение публичного
   контекста, явная политика Manager selector и tenant-aware offers/цены.
+- Подготовительное разделение Order выполняется отдельными совместимыми
+  релизами по плану [`order-domain-refactor.md`](order-domain-refactor.md).
 
 ## Зачем это делаем
